refactor(camera): split manager.go god-object into responsibility files (#225)
manager.go was 959 LOC mixing 7 distinct responsibilities. Split along
those boundaries into focused files (mirrors the existing partial split:
crud.go, timelapse.go, rediscovery.go, recorder_factory.go, etc.):

- manager.go (207 LOC, was 959): struct definition + constructor + types
  (CameraUpdate, RelayManager interface)
- lifecycle.go (427 LOC): Start/Stop + per-camera manual controls
  (StartCamera/StopCamera/RestartRecorder/SetProtocolEnabled) + the
  getRecorderHub and classifyError helpers
- hub.go (90 LOC): StreamHub lifecycle (GetOrCreateHub) + metrics wiring
- transcode.go (56 LOC): SetTranscodeManager + EnqueueTranscode
- stream_keys.go (92 LOC): RTMP/SRT stream-key resolution
  (RTMPKeyMap/ResolveStreamKey/SRTStreamConfigs/GetStreamURL)
- config_accessors.go (42 LOC): lock-free config/manager reads
  (CameraCount/GetCameraConfig/GetIngestRecorder/GetTimelapseMergeMgr)
- status.go: gained the health-manager integration methods
  (SetHealthManager/statusSnapshot/markStartFailed/clearStartFailed) that
  pair with its existing status accessors
- onvif_client.go: gained autoPopulateSnapshotURL (ONVIF snapshot helper)

No behavioral change: all exported signatures preserved, all methods
moved verbatim with original comments. The concurrency model (lock-free
COW snapshot registry in registry.go) is untouched.

Verified: go build ./... passes; go test -race ./internal/camera/... passes
(50s); coverage 47.6% -> 48.8%; go vet clean.

Closes #225
